% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGAMUN.R
\name{mogamun.load.data}
\alias{mogamun.load.data}
\title{mogamun.load.data}
\usage{
mogamun.load.data(
  DifferentialExpressionPath,
  NodesScoresPath,
  NetworkLayersDir,
  Layers
)
}
\arguments{
\item{DifferentialExpressionPath}{- full path to the differential expression results file (in CSV format). This file must contain It must contain at least the columns "gene" with the gene names, and ("PValue" or "FDR"). It can also contain "logFC"}

\item{NodesScoresPath}{- full path to the file containing the nodes scores (in CSV format). It must contain two columns: "gene" and "nodescore". NOTE. If no file is available, it wll be automatically generated}

\item{NetworkLayersDir}{- path of the folder that contains the networks that will be the layers of the multiplex}

\item{Layers}{- string of numbers, where the numbers correspond to the first character of the name of the network files}
}
\value{
None
}
\description{
Load the data to process
}
\examples{
mogamun.init()
mogamun.load.data(DifferentialExpressionPath =
                   "~/My_Experiment/Differential_Expression_Results.csv",
        NodesScoresPath = "~/My_Experiment/Genes_With_Nodes_Scores.csv",
        NetworkLayersDir = "~/My_Experiment/Layers_Multiplex/",
        Layers = "123")

}
