% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGAMUN.R
\name{mogamun.load.data}
\alias{mogamun.load.data}
\title{mogamun.load.data}
\usage{
mogamun.load.data(
  EvolutionParameters,
  DifferentialExpressionPath,
  NodesScoresPath,
  NetworkLayersDir,
  Layers
)
}
\arguments{
\item{EvolutionParameters}{evolution paramenters returned by mogamun.init()}

\item{DifferentialExpressionPath}{full path to the differential expression results file (in CSV format). This file must contain It must contain at least the columns "gene" with the gene names, and ("PValue" or "FDR"). It can also contain "logFC"}

\item{NodesScoresPath}{full path to an existing CSV file containing the nodes scores (columns "gene" and "nodescore"). NOTE. If the file does not exist, MOGAMUN will generate it in the provided path with the specified name}

\item{NetworkLayersDir}{path of the folder that contains the networks that will be the layers of the multiplex}

\item{Layers}{string of numbers, where the numbers correspond to the first character of the name of the network files}
}
\value{
List with the data to process
}
\description{
Load the data to process
}
\examples{
EvolutionParameters <- mogamun.init()
DEGPath <- system.file("ExampleFiles/DE/Banerji2017.csv", package = "MOGAMUN")
NodesScoresPath <- system.file("ExampleFiles/DE/Banerji2017_NodesScore.csv", package = "MOGAMUN")
LayersPath <- system.file("ExampleFiles/LayersMultiplex/", package = "MOGAMUN")
LoadedData <- mogamun.load.data(EvolutionParameters = EvolutionParameters,
                  DifferentialExpressionPath = DEGPath,
                  NodesScoresPath = NodesScoresPath,
                  NetworkLayersDir = LayersPath,
                  Layers = "123")

}
