% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGAMUN.R
\name{mogamun.run}
\alias{mogamun.run}
\title{mogamun.run}
\usage{
mogamun.run(resultsDir = ".")
}
\arguments{
\item{resultsDir}{outputs the results in the specified folder}
}
\value{
None
}
\description{
Run the algo
}
\examples{
mogamun.init(Generations=0)
mogamun.load.data(DifferentialExpressionPath =
                   "~/My_Experiment/Differential_Expression_Results.csv",
        NodesScoresPath = "~/My_Experiment/Genes_With_Nodes_Scores.csv",
        NetworkLayersDir = "~/My_Experiment/Layers_Multiplex/",
        Layers = "123")
mogamun.run(resultsDir = '~/My_Experiment/MOGAMUN_Results/')

}