% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGAMUN.R
\name{mogamun.run}
\alias{mogamun.run}
\title{mogamun.run}
\usage{
mogamun.run(resultsDir = ".")
}
\arguments{
\item{resultsDir}{outputs the results in the specified folder}
}
\value{
None
}
\description{
Run the algo
}
\examples{
mogamun.init(Generations=0)

dgepath <- system.file("ExampleFiles/DifferentialExpressionData/Banerji2017.csv", package = "MOGAMUN")
nodescorepath <- system.file("ExampleFiles/DifferentialExpressionData/Banerji2017_NodesScore.csv", package = "MOGAMUN")
layerspath <- system.file("ExampleFiles/LayersMultiplex/", package = "MOGAMUN")
mogamun.load.data(DifferentialExpressionPath = dgepath,
                  NodesScoresPath = nodescorepath,
                  NetworkLayersDir = layerspath,
                  Layers = "123")
mogamun.run(resultsDir = tempdir())

}
